Change Foreign Key relationship for ResourceBase and Thumbnail
Prevents duplicate thumbnails when updatelayers is run

Restores lost logic in updatelayers gs_slurp from old bulk merge

+++ geonode/base/tests.py
@@ -8,9 +8,11 @@ class ThumbnailTests(TestCase):
         self.rb = ResourceBase.objects.create()
 
     def tearDown(self):
-        t = self.rb.thumbnail
-        if t:
-            t.delete()
+        # if the thumbnail hasn't already been removed:
+        if self.rb.thumbnail_set.exists():
+            t = self.rb.thumbnail_set.get()
+            if t:
+                t.delete()
 
     def test_initial_behavior(self):
         self.assertFalse(self.rb.has_thumbnail())
